split up server.py in multiple files
